Added gallery query
Use authentication token to select all galleries the current user is a member of.
