+++ frontend/src/App.js
@@ -1,21 +1,26 @@
 import './style/app.css';
 import './style/settings.css';
 import './style/style.css';
-import React, { useState, useCallback } from 'react';
+import React, { useState } from 'react';
 import Login from './pages/Login.js';
 import Main from './pages/Main.js';
 function App() {
   
   const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [userID, setUserID] = useState("false");
   const [userData, setUserData] = useState(null);
   const [galleries, setGalleries] = useState([]);
   //const [token, setToken] = useState('');
   
+  const setCookie = (name, value, days) => {
+    const date = new Date();
+    date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000); // Expiration in days
+    const expires = `expires=${date.toUTCString()}`;
+    document.cookie = `${name}=${value};${expires};path=/`; // Set cookie
+  };
+  
   const handleLogin = async (email, password) => {
     // Here, you can add authentication logic (API call or checking credentials)
     // For now, just set it to true to simulate successful login
-    
     console.log("DEBUG: handleLogin called with:", email, password);
     //setIsLoggedIn(true);
         try {
@@ -29,9 +34,9 @@ function App() {
           if (!loginResponse.ok) throw new Error('Login failed');
           const { token } = await loginResponse.json();
           
-          //console.log("Login Success:", JSON.stringify({ email, password }));   
+          console.log("Login Success:", JSON.stringify({ email, password }));   
 
-
+          
           // Step 2: Get user info using token
           const userResponse = await fetch('/api/auth/me', {
             method: 'GET',
@@ -41,6 +46,7 @@ function App() {
           const userInfo = await userResponse.json();
           setUserData(userInfo); // Save user info
           //console.log("userResponse Success:", userInfo);   
+
           // Step 3: Fetch galleries using user ID
           const galleriesResponse = await fetch(`/api/user/galleries`, {
             method: 'GET',
@@ -52,6 +58,9 @@ function App() {
           
           setGalleries(galleryData); // Save galleries
           setIsLoggedIn(true); // Mark as logged in 
+          // Example usage after login
+          setCookie('authToken', token, 1); // Cookie valid for 7 days
+          
           
         } catch (error) {
           console.error("Error during login:", error);
