+++ frontend/src/style/app.css
@@ -1,5 +1,6 @@
 :root{
   --color-accent: #c9ffed;
+  --color-bar: #76bea6;
   --color-bkg: azure;
   --user-sidebar-length: 3.5vw;
   --user-dm-length: 2.5vw;
@@ -45,6 +46,7 @@ body{
   backdrop-filter: brightness(90%) sepia(20%) !important;
   background-repeat: no-repeat !important;
   width: 100% !important;
+  overflow-x: hidden;
 }
 .carousel-caption{
   width: 100% !important;
@@ -54,6 +56,7 @@ body{
 #gallery-sidebar-resizable{
   height: 100vh !important;
   background-color: var(--color-accent) !important;
+  background: linear-gradient(to right, var(--color-accent), var(--color-bar)) !important;
   overflow: hidden !important;
   /*position: absolute !important;
   left: 0;
@@ -107,7 +110,8 @@ body{
   background-size: cover !important;*/
   opacity: 100%;
   height: 100%;
-  width: 100vw;
+  width: auto;
+  position: relative;
 }
 .gallery-pane{
   width: 100% !important;
@@ -132,7 +136,8 @@ body{
   }
 }
 #sidebar-dm{
-  background-color: white;
+  /*background-color: white;*/
+  background: linear-gradient(to right, var(--color-bar), white);
   padding-left: 12px;
   max-height: 100% !important;
   height: 100vh !important;
@@ -140,7 +145,8 @@ body{
 }
 #sidebar-dm-options{
   font-weight: bolder;
-  background-color: white;
+  /*background-color: white;
+  background: linear-gradient(to right, var(--color-bar), white);*/
   text-align: center;
   width: 300px;
   height: fit-content;
@@ -157,7 +163,8 @@ body{
   position: relative !important;
 }
 #sidebar-channels {
-  background-color: white;
+  /*background-color: white;*/
+  background: linear-gradient(to right, var(--color-bar), rgba(154, 218, 197, 0.95), rgba(154, 218, 197, 0.80), rgba(255, 255, 255, 0.65));
   /*float: left;
   margin-left: 20px;*/
   height: 100vh !important;
